--- api/src/Controllers/getRecipesByName.js
@@ -1,13 +1,0 @@
-const axios = require('axios');
-
-const getRecipesByName = async () => {
-    try {
-        
-    } catch (error) {
-        throw new Error("Ups! We got a problem.")  
-    }
-};
-
-module.exports = {
-	getRecipesByName,
-};
