+++ api/src/db.js
@@ -44,8 +44,7 @@ const { Recipe, Diet} = sequelize.models;
 Recipe.belongsToMany(Diet, {through: "Recipe_Diet"});
 Diet.belongsToMany(Recipe, {through: "Recipe_Diet"})
 // Product.hasMany(Reviews);
-Recipe.belongsToMany(Diet, {through: "Recipe_Diet"})
-Diet.belongsToMany(Recipe, {through: "Recipe_Diet"})
+
 
 module.exports = {
   ...sequelize.models, // para poder importar los modelos así: const { Product, User } = require('./db.js');
