+++ api/src/helpers/infoMap.js
